Add packaging and docset feed
Docset now includes icons extracted from Open Group website favicons.
Should probably use something else.

Makefile now includes an `archive` target to build a tarball to be used
for the feed and Dash User Contributions package.

Includes feed XML file. The built tarball is being left in the
repository so the feed is self-hosting. In the future the tarball should
be moved to GitHub releases? Feed won't be necessary if the package is
accepted into the Dash User Contributions repository.

Includes docset.json for use in the Dash User Contributions package.